+++ src/NServiceBus.Metrics/ProbeContext.cs
@@ -1,0 +1,29 @@
+﻿namespace NServiceBus
+{
+    using System.Collections.Generic;
+
+    /// <summary>
+    /// Stores available probes
+    /// </summary>
+    public class ProbeContext
+    {
+        /// <summary>
+        /// Creates <see cref="ProbeContext"/>.
+        /// </summary>
+        public ProbeContext(IDurationProbe[] durations, ISignalProbe[] signals)
+        {
+            Durations = durations;
+            Signals = signals;
+        }
+
+        /// <summary>
+        /// Duration type probes.
+        /// </summary>
+        public IReadOnlyCollection<IDurationProbe> Durations { get; }
+
+        /// <summary>
+        /// Signal type probes.
+        /// </summary>
+        public IReadOnlyCollection<ISignalProbe> Signals { get; }
+    }
+}
