+++ src/NServiceBus.Metrics/TaskExtensions.cs
@@ -1,0 +1,9 @@
+﻿namespace NServiceBus.Metrics
+{
+    using System.Threading.Tasks;
+
+    class TaskExtensions
+    {
+        public static Task Completed => Task.FromResult(0);
+    }
+}
